Add module.h, update RSBE01.lst and other headers to support st_otrain

+++ Brawl/Include/gf/gf_archive.h
@@ -44,7 +44,7 @@ public:
     gfArchive();
     gfArchive(HeapType heap, int* unk1);
     virtual void notify(int unk1, gfFileIORequest* unk2);
-    void* getData(ARCNodeType nodeType, int fileIndex, u32 endian);
+    void* getData(ARCNodeType nodeType, int fileIndex, u16 endian);
     void* getData(int groupID, ARCNodeType nodeType, int fileIndex, u32 endian);
     void* getData(ARCNodeType nodeType, int fileIndex, int* nodeSize, u32 endian);
     void* getData(int groupID, ARCNodeType nodeType, int fileIndex, int* nodeSize, u32 endian);
